+++ polyfill/debug.js
@@ -1,0 +1,12 @@
+
+Object.defineProperty(exports, "__esModule", { value: true });
+
+function log() {
+    // ignore
+}
+
+function createDebug(namespace) {
+    return log;
+}
+
+exports.default = createDebug;